Fixing csv dependency.
When I started this project I intended to make a full csv list to reference, but league is a dynamic game that needs constant updating, so I decided to call from the ddragon API by pulling the JSON data, parsing it, and receiving the cooldown information. It is still a work in progress, but a step in the right direction.